update licence and cleanup code
