+++ .gitignore
@@ -100,3 +100,5 @@ browserstack.err
 cache/
 custom/**/
 .sass-cache/
+
+.php_cs.cache
