+++ include/SugarLogger/LoggerManager.php
@@ -1,11 +1,11 @@
 <?php
-if(!defined('sugarEntry') || !sugarEntry) die('Not A Valid Entry Point');
-/*********************************************************************************
+/**
+ *
  * SugarCRM Community Edition is a customer relationship management program developed by
  * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
-
- * SuiteCRM is an extension to SugarCRM Community Edition developed by Salesagility Ltd.
- * Copyright (C) 2011 - 2014 Salesagility Ltd.
+ *
+ * SuiteCRM is an extension to SugarCRM Community Edition developed by SalesAgility Ltd.
+ * Copyright (C) 2011 - 2018 SalesAgility Ltd.
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License version 3 as published by the
@@ -16,7 +16,7 @@ if(!defined('sugarEntry') || !sugarEntry) die('Not A Valid Entry Point');
  *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
+ * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
  * details.
  *
  * You should have received a copy of the GNU Affero General Public License along with
@@ -34,9 +34,14 @@ if(!defined('sugarEntry') || !sugarEntry) die('Not A Valid Entry Point');
  * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
  * these Appropriate Legal Notices must retain the display of the "Powered by
  * SugarCRM" logo and "Supercharged by SuiteCRM" logo. If the display of the logos is not
- * reasonably feasible for  technical reasons, the Appropriate Legal Notices must
- * display the words  "Powered by SugarCRM" and "Supercharged by SuiteCRM".
- ********************************************************************************/
+ * reasonably feasible for technical reasons, the Appropriate Legal Notices must
+ * display the words "Powered by SugarCRM" and "Supercharged by SuiteCRM".
+ */
+
+if (!defined('sugarEntry') || !sugarEntry) {
+    die('Not A Valid Entry Point');
+}
+
 
 /**
  * Log management
@@ -84,37 +89,37 @@ class LoggerManager
 		    $this->_findAvailableLoggers();
 	}
 
-	/**
-	 * Overloaded method that handles the logging requests.
-	 *
-	 * @param string $method
-	 * @param string $message - also handles array as parameter, though that is deprecated.
-	 */
- 	public function __call(
- 	    $method,
- 	    $message
- 	    )
- 	{
-                global $sugar_config;
+    /**
+     * Overloaded method that handles the logging requests.
+     *
+     * @param string $method
+     * @param string $message - also handles array as parameter, though that is deprecated.
+     */
+    public function __call(
+        $method,
+        $message
+        )
+    {
+        global $sugar_config;
 
-                // show calls stack trace
+        // show calls stack trace
 
-                if(isset($sugar_config['show_log_trace']) && $sugar_config['show_log_trace']) {
-                        try {
-                                throw new Exception();
-                        } catch (Exception $e) {
-                                $trace = "\n" . $e->getTraceAsString() . "\n";
-                                if(is_string($message)) {
-                                        $message .= $trace;
-                                } else {
-                                        if(is_array($message)) {
-                                                foreach($message as &$msg) {
-                                                        $msg .= $trace;
-                                                }
-                                        }
-                                }
+        if (isset($sugar_config['show_log_trace']) && $sugar_config['show_log_trace']) {
+            try {
+                throw new Exception();
+            } catch (Exception $e) {
+                $trace = "\n" . $e->getTraceAsString() . "\n";
+                if (is_string($message)) {
+                    $message .= $trace;
+                } else {
+                    if (is_array($message)) {
+                        foreach ($message as &$msg) {
+                            $msg .= $trace;
                         }
+                    }
                 }
+            }
+        }
 
         if ( !isset(self::$_levelMapping[$method]) )
             $method = $this->_level;
