+++ modules/Contacts/AOPContactUtils.php
@@ -5,7 +5,7 @@
  * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
  *
  * SuiteCRM is an extension to SugarCRM Community Edition developed by SalesAgility Ltd.
- * Copyright (C) 2011 - 2017 SalesAgility Ltd.
+ * Copyright (C) 2011 - 2018 SalesAgility Ltd.
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License version 3 as published by the
@@ -42,14 +42,9 @@ if (!defined('sugarEntry') || !sugarEntry) {
     die('Not A Valid Entry Point');
 }
 
-/**
- * class AOPContactUtilsException
- */
-class AOPContactUtilsException extends Exception {
 
-    const UNABLE_READ_PORTAL_VERSION = 100;
-
-}
+require_once __DIR__ . '/../AOP_Case_Updates/util.php';
+require_once __DIR__ . '/AOPContactUtilsException.php';
 
 /**
  * Class AOPContactUtils
