+++ modules/Contacts/AOPContactUtilsException.php
@@ -1,0 +1,12 @@
+<?php
+
+
+/**
+ * class AOPContactUtilsException
+ */
+class AOPContactUtilsException extends Exception {
+
+    const UNABLE_READ_PORTAL_VERSION = 100;
+
+}
+
