+++ modules/Contacts/JAccountMissingException.php
@@ -1,0 +1,5 @@
+<?php
+
+
+class JAccountMissingException extends Exception {}
+
