+++ modules/Contacts/disablePortalUser.php
@@ -1,6 +1,46 @@
 <?php
 /**
  *
+ * SugarCRM Community Edition is a customer relationship management program developed by
+ * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
+ *
+ * SuiteCRM is an extension to SugarCRM Community Edition developed by SalesAgility Ltd.
+ * Copyright (C) 2011 - 2018 SalesAgility Ltd.
+ *
+ * This program is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Affero General Public License version 3 as published by the
+ * Free Software Foundation with the addition of the following permission added
+ * to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
+ * IN WHICH THE COPYRIGHT IS OWNED BY SUGARCRM, SUGARCRM DISCLAIMS THE WARRANTY
+ * OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU Affero General Public License along with
+ * this program; if not, see http://www.gnu.org/licenses or write to the Free
+ * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+ * 02110-1301 USA.
+ *
+ * You can contact SugarCRM, Inc. headquarters at 10050 North Wolfe Road,
+ * SW2-130, Cupertino, CA 95014, USA. or at email address contact@sugarcrm.com.
+ *
+ * The interactive user interfaces in modified source and object code versions
+ * of this program must display Appropriate Legal Notices, as required under
+ * Section 5 of the GNU Affero General Public License version 3.
+ *
+ * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
+ * these Appropriate Legal Notices must retain the display of the "Powered by
+ * SugarCRM" logo and "Supercharged by SuiteCRM" logo. If the display of the logos is not
+ * reasonably feasible for technical reasons, the Appropriate Legal Notices must
+ * display the words "Powered by SugarCRM" and "Supercharged by SuiteCRM".
+ */
+
+
+/**
+ *
  * @package Advanced OpenPortal
  * @copyright SalesAgility Ltd http://www.salesagility.com
  *
@@ -21,27 +61,22 @@
  *
  * @author Salesagility Ltd <support@salesagility.com>
  */
-if(!defined('sugarEntry'))define('sugarEntry', true);
 
-include_once 'modules/Contacts/AOPContactUtils.php';
-
-class JAccountMissingException extends Exception {}
-
-require_once 'modules/AOP_Case_Updates/util.php';
-
-function url_exists($url)
-{
-    if (!$fp = curl_init($url)) return false;
-    return true;
+if (!defined('sugarEntry') || !sugarEntry) {
+    die('Not A Valid Entry Point');
 }
 
+include_once __DIR__ . '/AOPContactUtils.php';
+include_once __DIR__ . '/JAccountMissingException.php';
+
+require_once __DIR__ . '/../AOP_Case_Updates/util.php';
 
 if (!isAOPEnabled()) {
     return;
 }
 global $sugar_config, $mod_strings;
 
-require_once('modules/Contacts/Contact.php');
+require_once __DIR__ . '/Contact.php';
 
 $bean = new Contact();
 $bean->retrieve($_REQUEST['record']);
