+++ modules/Contacts/metadata/detailviewdefs.php
@@ -38,6 +38,11 @@
  * display the words "Powered by SugarCRM" and "Supercharged by SuiteCRM".
  */
 
+if (!defined('sugarEntry') || !sugarEntry) {
+    die('Not A Valid Entry Point');
+}
+
+
 $viewdefs ['Contacts'] = 
 array (
   'DetailView' =>
