+++ modules/Contacts/updatePortal.php
@@ -5,7 +5,7 @@
  * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
  *
  * SuiteCRM is an extension to SugarCRM Community Edition developed by SalesAgility Ltd.
- * Copyright (C) 2011 - 2016 SalesAgility Ltd.
+ * Copyright (C) 2011 - 2018 SalesAgility Ltd.
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License version 3 as published by the
@@ -16,7 +16,7 @@
  *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
+ * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
  * details.
  *
  * You should have received a copy of the GNU Affero General Public License along with
@@ -34,26 +34,35 @@
  * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
  * these Appropriate Legal Notices must retain the display of the "Powered by
  * SugarCRM" logo and "Supercharged by SuiteCRM" logo. If the display of the logos is not
- * reasonably feasible for  technical reasons, the Appropriate Legal Notices must
- * display the words  "Powered by SugarCRM" and "Supercharged by SuiteCRM".
+ * reasonably feasible for technical reasons, the Appropriate Legal Notices must
+ * display the words "Powered by SugarCRM" and "Supercharged by SuiteCRM".
  */
 
 if (!defined('sugarEntry') || !sugarEntry) {
     die('Not A Valid Entry Point');
 }
 
-require_once 'modules/AOP_Case_Updates/util.php';
 
+require_once __DIR__ . '/../AOP_Case_Updates/util.php';
+
+/**
+ * class UpdatePortal
+ */
 class updatePortal
 {    
     
+    /**
+     *
+     * @var bool 
+     */
     private static $pwdEmailSent = false;
 
     /**
-     * @param $bean
-     * @return bool
-     * @throws JAccountException
-
+     * 
+     * @global array $sugar_config
+     * @param Account|Contact $bean
+     * @return boolean
+     * @todo parameter needs an interface type hinting to using joomla_account_access field with a getter
      */
     public function updateUser($bean)
     {
@@ -117,13 +126,13 @@ class updatePortal
     }
 
     /**
-     * @param $emailTo
-     * @param $emailSubject
-     * @param $emailBody
-     * @param $altEmailBody
+     * 
+     * @param string $emailTo
+     * @param string $emailSubject
+     * @param string $emailBody
+     * @param string $altEmailBody
      * @param SugarBean|null $relatedBean
      * @return bool
-     * @throws phpmailerException
      */
     public function sendEmail($emailTo, $emailSubject, $emailBody, $altEmailBody, SugarBean $relatedBean = null)
     {
@@ -146,7 +155,7 @@ class updatePortal
         $mail->addAddress($emailTo);
 
         //now create email
-        $ok = @$mail->send();
+        $ok = $mail->send();
         if ($ok) {
 
             $emailObj->to_addrs_names = $emailTo;
