+++ modules/JAccount/JAccount.php
@@ -5,7 +5,7 @@
  * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
  *
  * SuiteCRM is an extension to SugarCRM Community Edition developed by SalesAgility Ltd.
- * Copyright (C) 2011 - 2017 SalesAgility Ltd.
+ * Copyright (C) 2011 - 2018 SalesAgility Ltd.
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License version 3 as published by the
@@ -16,7 +16,7 @@
  *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
+ * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
  * details.
  *
  * You should have received a copy of the GNU Affero General Public License along with
@@ -34,43 +34,162 @@
  * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
  * these Appropriate Legal Notices must retain the display of the "Powered by
  * SugarCRM" logo and "Supercharged by SuiteCRM" logo. If the display of the logos is not
- * reasonably feasible for  technical reasons, the Appropriate Legal Notices must
- * display the words  "Powered by SugarCRM" and "Supercharged by SuiteCRM".
+ * reasonably feasible for technical reasons, the Appropriate Legal Notices must
+ * display the words "Powered by SugarCRM" and "Supercharged by SuiteCRM".
  */
 
-include_once 'modules/JAccount/JAccountException.php';
-include_once 'modules/JAccount/JAccountInvalidUserDataException.php';
+if (!defined('sugarEntry') || !sugarEntry) {
+    die('Not A Valid Entry Point');
+}
 
 
+include_once __DIR__ . '/JAccountException.php';
+include_once __DIR__ . '/JAccountInvalidUserDataException.php';
+
+
+/**
+ * class JAccount
+ */
 class JAccount extends Basic
 {
+    /**
+     *
+     * @var bool 
+     */
     public $new_schema = true;
+    
+    /**
+     *
+     * @var string 
+     */
     public $module_dir = 'JAccount';
+    
+    /**
+     *
+     * @var string 
+     */
     public $object_name = 'JAccount';
+    
+    /**
+     *
+     * @var string 
+     */
     public $table_name = 'jaccount';
+    
+    /**
+     *
+     * @var bool 
+     */
     public $importable = false;
 
+    /**
+     *
+     * @var string 
+     */
     public $id;
+    
+    /**
+     *
+     * @var string 
+     */
     public $name;
+    
+    /**
+     *
+     * @var string 
+     */
     public $date_entered;
+    
+    /**
+     *
+     * @var string 
+     */
     public $date_modified;
+    
+    /**
+     *
+     * @var string 
+     */
     public $modified_user_id;
+    
+    /**
+     *
+     * @var string 
+     */
     public $modified_by_name;
+    
+    /**
+     *
+     * @var string 
+     */
     public $created_by;
+    
+    /**
+     *
+     * @var string 
+     */
     public $created_by_name;
+    
+    /**
+     *
+     * @var string 
+     */
     public $description;
+    
+    /**
+     *
+     * @var int 
+     */
     public $deleted;
+    
+    /**
+     *
+     * @var string 
+     */
     public $created_by_link;
+    
+    /**
+     *
+     * @var string 
+     */
     public $modified_user_link;
+    
+    /**
+     *
+     * @var string 
+     */
     public $assigned_user_id;
+    
+    /**
+     *
+     * @var string 
+     */
     public $assigned_user_name;
+    
+    /**
+     *
+     * @var string 
+     */
     public $assigned_user_link;
+    
+    /**
+     *
+     * @var SecurityGroups
+     */
     public $SecurityGroups;
-
+    
+    /**
+     * constructor
+     */
     public function __construct() {
         parent::__construct();
     }
 	
+    /**
+     * 
+     * @param string $interface
+     * @return boolean
+     */
     public function bean_implements($interface)
     {
         switch($interface)
@@ -82,6 +201,13 @@ class JAccount extends Basic
         return false;
     }
 
+    /**
+     * 
+     * @param int|string|null $id
+     * @param bool|null $encode
+     * @param bool|null $deleted
+     * @return SugarBean|null
+     */
     public function retrieve($id = -1, $encode = true, $deleted = true) {
         $ret = parent::retrieve($id, $encode, $deleted);
         if($ret) {
@@ -94,6 +220,13 @@ class JAccount extends Basic
         return $ret;
     }
 
+    /**
+     * 
+     * @param bool $check_notify
+     * @return string ID
+     * @throws JAccountInvalidUserDataException
+     * @throws JAccountException
+     */
     public function save($check_notify = false) {
         if(!isset($this->name) || !$this->name || !isset($this->email1) || !$this->email1) {
             throw new JAccountInvalidUserDataException("Trying to save a JAccount without name and/or any email address: (name: {$this->name}, email1: {$this->email1})");
@@ -117,7 +250,7 @@ class JAccount extends Basic
 
             // todo: (scrm-436) but now, the Exception without type could be everything..
 
-            // DBManager drop an exception when duplicate index error on insert so de bean already there, retrieve it..
+            // DBManager drop an exception when duplicate index error on insert so a bean already there, retrieve it..
 
             $ret = $this->retrieve_by_string_fields(array(
                 'contact_id' => $contact->id,
@@ -139,7 +272,7 @@ class JAccount extends Basic
      * and handle the errors correctly
      * (return false on error)
      *
-     * @param $request $_REQUEST
+     * @param array $request $_REQUEST
      * @return bool|JAccount
      * @throws JAccountException
      */
